+++ normalize/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - add `identify` method
 - export utils seperately
+- add `reachableDataIds` method
 
 ### Changed